fix: change invoice number query to use Long instead of Integer
The invoice number format (INV202601090001) produces a 12-digit number
when extracting the numeric part, which exceeds Integer.MAX_VALUE.
Changed CAST to long and return type to Long to handle larger values.

+++ shina-magazin-api/src/main/java/uz/shinamagazin/api/service/SaleService.java
@@ -238,8 +238,8 @@ public class SaleService {
 
     private String generateInvoiceNumber() {
         String prefix = "INV" + LocalDate.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"));
-        Integer maxNumber = saleRepository.findMaxInvoiceNumber(prefix);
-        int nextNumber = (maxNumber != null ? maxNumber : 0) + 1;
+        Long maxNumber = saleRepository.findMaxInvoiceNumber(prefix);
+        long nextNumber = (maxNumber != null ? maxNumber : 0L) + 1;
         return prefix + String.format("%04d", nextNumber);
     }
